Consolidate duplicated GPU logic into new GPUArrays extension
