+++ ext/TensorKitAMDGPUExt/TensorKitAMDGPUExt.jl
@@ -1,7 +1,6 @@
 module TensorKitAMDGPUExt
 
 using AMDGPU, AMDGPU.rocBLAS, AMDGPU.rocSOLVER, LinearAlgebra
-using AMDGPU: @allowscalar
 import AMDGPU: rand as rocrand, rand! as rocrand!, randn as rocrandn, randn! as rocrandn!
 
 using TensorKit
@@ -9,7 +8,7 @@ using TensorKit.Factorizations
 using Strided
 using MatrixAlgebraKit
 using MatrixAlgebraKit: AbstractAlgorithm
-using TensorKit: SectorDict, tensormaptype, scalar, similarstoragetype, AdjointTensorMap, scalartype, project_symmetric_and_check
+using TensorKit: SectorDict, tensormaptype, scalar, similarstoragetype, AdjointTensorMap, scalartype
 import TensorKit: randisometry
 using Base: rand, randn
