+++ ext/TensorKitAMDGPUExt/roctensormap.jl
@@ -7,16 +7,6 @@ function ROCTensorMap(t::TensorMap{T, S, N₁, N₂, A}) where {T, S, N₁, N₂
     return ROCTensorMap{T, S, N₁, N₂}(ROCArray{T}(t.data), space(t))
 end
 
-# project_symmetric! doesn't yet work for GPU types, so do this on the host, then copy
-function TensorKit.project_symmetric_and_check(::Type{T}, ::Type{A}, data::AbstractArray, V::TensorMapSpace; tol = sqrt(eps(real(float(eltype(data)))))) where {T, A <: ROCVector{T}}
-    h_t = TensorKit.TensorMapWithStorage{T, Vector{T}}(undef, V)
-    h_t = TensorKit.project_symmetric!(h_t, Array(data))
-    # verify result
-    isapprox(Array(reshape(data, dims(h_t))), convert(Array, h_t); atol = tol) ||
-        throw(ArgumentError("Data has non-zero elements at incompatible positions"))
-    return TensorKit.TensorMapWithStorage{T, A}(A(h_t.data), V)
-end
-
 for (fname, felt) in ((:zeros, :zero), (:ones, :one))
     @eval begin
         function AMDGPU.$fname(
@@ -90,13 +80,6 @@ for randfun in (:rocrand, :rocrandn)
             return t
         end
     end
-end
-
-# Scalar implementation
-#-----------------------
-function TensorKit.scalar(t::ROCTensorMap{T, S, 0, 0}) where {T, S}
-    inds = findall(!iszero, t.data)
-    return isempty(inds) ? zero(scalartype(t)) : @allowscalar @inbounds t.data[only(inds)]
 end
 
 function Base.convert(
