+++ ext/TensorKitCUDAExt/TensorKitCUDAExt.jl
@@ -1,10 +1,8 @@
 module TensorKitCUDAExt
 
 using CUDA, CUDA.cuBLAS, CUDA.cuSOLVER, CUDA.cuRAND, LinearAlgebra
-using CUDA: @allowscalar
 import CUDA.cuRAND: rand as curand, rand! as curand!, randn as curandn, randn! as curandn!
 using Strided: StridedViews
-using CUDA.CUDACore.KernelAbstractions: @kernel, @index, get_backend
 
 using Adapt: Adapt
 
@@ -20,20 +18,5 @@ using TensorKit: MatrixAlgebraKit
 using Random
 
 include("cutensormap.jl")
-include("truncation.jl")
-
-function TensorKit.fill_braidingsubblock!(data::TD, val) where {T, TD <: Union{<:CuMatrix{T}, <:StridedViews.StridedView{T, 4, <:CuArray{T}}}}
-    # COV_EXCL_START
-    # kernels are not reachable by coverage
-    @kernel function fill_subblock_kernel!(subblock, val)
-        idx = @index(Global, Cartesian)
-        idx_val = idx[1] == idx[4] && idx[2] == idx[3] ? val : zero(val)
-        @inbounds subblock[idx] = idx_val
-    end
-    # COV_EXCL_STOP
-    kernel = fill_subblock_kernel!(get_backend(data))
-    kernel(data, val; ndrange = size(data))
-    return data
-end
 
 end
