+++ ext/TensorKitCUDAExt/cutensormap.jl
@@ -7,16 +7,6 @@ function CuTensorMap(t::TensorMap{T, S, N₁, N₂, A}) where {T, S, N₁, N₂,
     return CuTensorMap{T, S, N₁, N₂}(CuArray{T}(t.data), space(t))
 end
 
-# project_symmetric! doesn't yet work for GPU types, so do this on the host, then copy
-function TensorKit.project_symmetric_and_check(::Type{T}, ::Type{A}, data::AbstractArray, V::TensorMapSpace; tol = sqrt(eps(real(float(eltype(data)))))) where {T, A <: CuVector{T}}
-    h_t = TensorKit.TensorMapWithStorage{T, Vector{T}}(undef, V)
-    h_t = TensorKit.project_symmetric!(h_t, Array(data))
-    # verify result
-    isapprox(Array(reshape(data, dims(h_t))), convert(Array, h_t); atol = tol) ||
-        throw(ArgumentError("Data has non-zero elements at incompatible positions"))
-    return TensorKit.TensorMapWithStorage{T, A}(A(h_t.data), V)
-end
-
 for (fname, felt) in ((:zeros, :zero), (:ones, :one))
     @eval begin
         function CUDA.$fname(
@@ -92,13 +82,6 @@ for randfun in (:curand, :curandn)
             return t
         end
     end
-end
-
-# Scalar implementation
-#-----------------------
-function TensorKit.scalar(t::CuTensorMap{T, S, 0, 0}) where {T, S}
-    inds = findall(!iszero, t.data)
-    return isempty(inds) ? zero(scalartype(t)) : @allowscalar @inbounds t.data[only(inds)]
 end
 
 function LinearAlgebra.isposdef(t::CuTensorMap)
