+++ test/Project.toml
@@ -36,7 +36,6 @@ AllocCheck = "0.2"
 ChainRulesTestUtils = "1"
 Combinatorics = "1"
 cuTENSOR = "6"
-GPUArrays = "11.3.1"
 JET = "0.9, 0.10, 0.11"
 ParallelTestRunner = "2"
 Test = "1"
